Added Defaults and Initialization
The initialize method binds "this" to get scoping to work correctly within the pagination object. We only need that in order to set pagination.collection to the parent collection. I'm still not sure this pattern is o.k. to use (I'd love feedback!). Also, since it was somewhat likely that the initialize method would be overwritten when this is actually used, I thought it better to make the initialize work in a separate function that can be called manually. Binding "this" seemed better than making the developer know to set "pagination.collection = this" in the initialize method they write when using Pageable.